requirements.txt
# Core BDD Framework
behave==1.2.6

# HTTP Client
requests==2.31.0

# Assertions & Testing
pytest==7.4.3

# Configuration Management
PyYAML==6.0.1

# Reporting
allure-behave==2.13.2

# Data Generation
Faker==20.1.0

# JSON Schema Validation
jsonschema==4.20.0

# Retry Mechanism
tenacity==8.2.3

# Environment Variables
python-dotenv==1.0.0

# Logging Enhancement
colorlog==6.8.0

# Parallel Execution
pytest-xdist==3.5.0

# Date/Time Utilities
python-dateutil==2.8.2
